refactor: Reorg identity provider to match structure (#406)
Refactor/reogranize identity provider code to match the structure of the
other providers. This improves modularization and help to extend
functionality.
