+++ src/identity/backends/sql/user/delete.rs
@@ -1,0 +1,72 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use sea_orm::DatabaseConnection;
+use sea_orm::entity::*;
+
+use crate::config::Config;
+use crate::db::entity::prelude::User as DbUser;
+use crate::identity::backends::sql::{IdentityDatabaseError, db_err};
+
+pub async fn delete<U: AsRef<str>>(
+    _conf: &Config,
+    db: &DatabaseConnection,
+    user_id: U,
+) -> Result<(), IdentityDatabaseError> {
+    let res = DbUser::delete_by_id(user_id.as_ref())
+        .exec(db)
+        .await
+        .map_err(|err| db_err(err, "deleting the user record"))?;
+    if res.rows_affected == 1 {
+        Ok(())
+    } else {
+        Err(IdentityDatabaseError::UserNotFound(
+            user_id.as_ref().to_string(),
+        ))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #![allow(clippy::derivable_impls)]
+
+    use sea_orm::{DatabaseBackend, MockDatabase, MockExecResult, Transaction};
+
+    use crate::identity::Config;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn test_delete() {
+        // Create MockDatabase with mock query results
+        let db = MockDatabase::new(DatabaseBackend::Postgres)
+            .append_exec_results([MockExecResult {
+                rows_affected: 1,
+                ..Default::default()
+            }])
+            .into_connection();
+        let config = Config::default();
+
+        delete(&config, &db, "id").await.unwrap();
+        // Checking transaction log
+        assert_eq!(
+            db.into_transaction_log(),
+            [Transaction::from_sql_and_values(
+                DatabaseBackend::Postgres,
+                r#"DELETE FROM "user" WHERE "user"."id" = $1"#,
+                ["id".into()]
+            ),]
+        );
+    }
+}
