+++ src/identity/mod.rs
@@ -13,365 +13,36 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use async_trait::async_trait;
-#[cfg(test)]
-use mockall::mock;
 use std::collections::HashSet;
 use uuid::Uuid;
 use webauthn_rs::prelude::{Passkey, PasskeyAuthentication, PasskeyRegistration};
 
 pub mod backends;
 pub mod error;
+#[cfg(test)]
+pub mod mock;
 pub mod password_hashing;
-pub(crate) mod types;
+pub mod types;
+#[cfg(test)]
+pub use mock::MockIdentityProvider;
 
 use crate::auth::AuthenticatedInfo;
 use crate::config::Config;
-use crate::identity::backends::sql::SqlBackend;
+use crate::identity::backends::{IdentityBackend, sql::SqlBackend};
 use crate::identity::error::IdentityProviderError;
 use crate::identity::types::{
-    Group, GroupCreate, GroupListParameters, IdentityBackend, UserCreate, UserListParameters,
+    Group, GroupCreate, GroupListParameters, UserCreate, UserListParameters,
     UserPasswordAuthRequest, UserResponse, WebauthnCredential,
 };
 use crate::keystone::ServiceState;
 use crate::plugin_manager::PluginManager;
 use crate::resource::{ResourceApi, error::ResourceProviderError};
 
+pub use types::IdentityApi;
+
 #[derive(Clone, Debug)]
 pub struct IdentityProvider {
     backend_driver: Box<dyn IdentityBackend>,
-}
-
-#[async_trait]
-pub trait IdentityApi: Send + Sync + Clone {
-    async fn authenticate_by_password(
-        &self,
-        state: &ServiceState,
-        auth: UserPasswordAuthRequest,
-    ) -> Result<AuthenticatedInfo, IdentityProviderError>;
-
-    async fn list_users(
-        &self,
-        state: &ServiceState,
-        params: &UserListParameters,
-    ) -> Result<impl IntoIterator<Item = UserResponse>, IdentityProviderError>;
-
-    async fn get_user<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<Option<UserResponse>, IdentityProviderError>;
-
-    async fn find_federated_user<'a>(
-        &self,
-        state: &ServiceState,
-        idp_id: &'a str,
-        unique_id: &'a str,
-    ) -> Result<Option<UserResponse>, IdentityProviderError>;
-
-    async fn create_user(
-        &self,
-        state: &ServiceState,
-        user: UserCreate,
-    ) -> Result<UserResponse, IdentityProviderError>;
-
-    async fn delete_user<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-
-    async fn list_groups(
-        &self,
-        state: &ServiceState,
-        params: &GroupListParameters,
-    ) -> Result<impl IntoIterator<Item = Group>, IdentityProviderError>;
-
-    async fn get_group<'a>(
-        &self,
-        state: &ServiceState,
-        group_id: &'a str,
-    ) -> Result<Option<Group>, IdentityProviderError>;
-
-    async fn create_group(
-        &self,
-        state: &ServiceState,
-        group: GroupCreate,
-    ) -> Result<Group, IdentityProviderError>;
-
-    async fn delete_group<'a>(
-        &self,
-        state: &ServiceState,
-        group_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// List groups the user is a member of.
-    async fn list_groups_of_user<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<impl IntoIterator<Item = Group>, IdentityProviderError>;
-
-    /// Add the user to the single group.
-    async fn add_user_to_group<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        group_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// Add user group memberships as specified by (uid, gid) tuples.
-    async fn add_users_to_groups<'a>(
-        &self,
-        state: &ServiceState,
-        memberships: Vec<(&'a str, &'a str)>,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// Remove the user from the single group.
-    async fn remove_user_from_group<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        group_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// Remove the user from specified groups.
-    async fn remove_user_from_groups<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        group_ids: HashSet<&'a str>,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// Set group memberships of the user.
-    async fn set_user_groups<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        group_ids: HashSet<&'a str>,
-    ) -> Result<(), IdentityProviderError>;
-
-    async fn list_user_webauthn_credentials<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<impl IntoIterator<Item = Passkey>, IdentityProviderError>;
-
-    /// Create passkey.
-    async fn create_user_webauthn_credential<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        passkey: &Passkey,
-        description: Option<&'a str>,
-    ) -> Result<WebauthnCredential, IdentityProviderError>;
-
-    async fn save_user_webauthn_credential_registration_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        state: PasskeyRegistration,
-    ) -> Result<(), IdentityProviderError>;
-
-    async fn save_user_webauthn_credential_authentication_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-        state: PasskeyAuthentication,
-    ) -> Result<(), IdentityProviderError>;
-
-    async fn get_user_webauthn_credential_registration_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<Option<PasskeyRegistration>, IdentityProviderError>;
-
-    async fn get_user_webauthn_credential_authentication_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<Option<PasskeyAuthentication>, IdentityProviderError>;
-
-    /// Delete passkey registration state of a user
-    async fn delete_user_webauthn_credential_registration_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-
-    /// Delete passkey registration state of a user
-    async fn delete_user_webauthn_credential_authentication_state<'a>(
-        &self,
-        state: &ServiceState,
-        user_id: &'a str,
-    ) -> Result<(), IdentityProviderError>;
-}
-
-#[cfg(test)]
-mock! {
-    pub IdentityProvider {
-        pub fn new(cfg: &Config, plugin_manager: &PluginManager) -> Result<Self, IdentityProviderError>;
-    }
-
-    #[async_trait]
-    impl IdentityApi for IdentityProvider {
-        async fn authenticate_by_password(
-            &self,
-            state: &ServiceState,
-            auth: UserPasswordAuthRequest,
-        ) -> Result<AuthenticatedInfo, IdentityProviderError>;
-
-        async fn list_users(
-            &self,
-            state: &ServiceState,
-            params: &UserListParameters,
-        ) -> Result<Vec<UserResponse>, IdentityProviderError>;
-
-        async fn get_user<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Option<UserResponse>, IdentityProviderError>;
-
-        async fn find_federated_user<'a>(
-            &self,
-            state: &ServiceState,
-            idp_id: &'a str,
-            unique_id: &'a str,
-        ) -> Result<Option<UserResponse>, IdentityProviderError>;
-
-        async fn create_user(
-            &self,
-            state: &ServiceState,
-            user: UserCreate,
-        ) -> Result<UserResponse, IdentityProviderError>;
-
-        async fn delete_user<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn list_groups(
-            &self,
-            state: &ServiceState,
-            params: &GroupListParameters,
-        ) -> Result<Vec<Group>, IdentityProviderError>;
-
-        async fn get_group<'a>(
-            &self,
-            state: &ServiceState,
-            group_id: &'a str,
-        ) -> Result<Option<Group>, IdentityProviderError>;
-
-        async fn create_group(
-            &self,
-            state: &ServiceState,
-            group: GroupCreate,
-        ) -> Result<Group, IdentityProviderError>;
-
-        async fn delete_group<'a>(
-            &self,
-            state: &ServiceState,
-            group_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn list_groups_of_user<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Vec<Group>, IdentityProviderError>;
-
-        async fn add_user_to_group<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            group_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn add_users_to_groups<'a>(
-            &self,
-            state: &ServiceState,
-            memberships: Vec<(&'a str, &'a str)>
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn remove_user_from_group<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            group_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn remove_user_from_groups<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            group_ids: HashSet<&'a str>,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn set_user_groups<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            group_ids: HashSet<&'a str>,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn list_user_webauthn_credentials<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Vec<Passkey>, IdentityProviderError>;
-
-        async fn create_user_webauthn_credential<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            passkey: &Passkey,
-            description: Option<&'a str>
-        ) -> Result<WebauthnCredential, IdentityProviderError>;
-
-        async fn save_user_webauthn_credential_registration_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            auth_state: PasskeyRegistration,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn save_user_webauthn_credential_authentication_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-            auth_state: PasskeyAuthentication,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn get_user_webauthn_credential_registration_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Option<PasskeyRegistration>, IdentityProviderError>;
-
-        async fn get_user_webauthn_credential_authentication_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<Option<PasskeyAuthentication>, IdentityProviderError>;
-
-        async fn delete_user_webauthn_credential_registration_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-
-        async fn delete_user_webauthn_credential_authentication_state<'a>(
-            &self,
-            state: &ServiceState,
-            user_id: &'a str,
-        ) -> Result<(), IdentityProviderError>;
-    }
-
-    impl Clone for IdentityProvider {
-        fn clone(&self) -> Self;
-    }
-
 }
 
 impl IdentityProvider {
