+++ src/identity/types.rs
@@ -18,43 +18,34 @@ pub mod group;
 pub mod user;
 
 use async_trait::async_trait;
-use dyn_clone::DynClone;
 use webauthn_rs::prelude::{Passkey, PasskeyAuthentication, PasskeyRegistration};
 
 use crate::auth::AuthenticatedInfo;
-use crate::config::Config;
 use crate::identity::IdentityProviderError;
 pub use crate::identity::types::group::{Group, GroupCreate, GroupListParameters};
 pub use crate::identity::types::user::*;
 use crate::keystone::ServiceState;
 
 #[async_trait]
-pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
-    /// Set config.
-    fn set_config(&mut self, config: Config);
-
-    /// Authenticate a user by a password.
+pub trait IdentityApi: Send + Sync + Clone {
     async fn authenticate_by_password(
         &self,
         state: &ServiceState,
         auth: UserPasswordAuthRequest,
     ) -> Result<AuthenticatedInfo, IdentityProviderError>;
 
-    /// List Users.
     async fn list_users(
         &self,
         state: &ServiceState,
         params: &UserListParameters,
-    ) -> Result<Vec<UserResponse>, IdentityProviderError>;
+    ) -> Result<impl IntoIterator<Item = UserResponse>, IdentityProviderError>;
 
-    /// Get single user by ID.
     async fn get_user<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<Option<UserResponse>, IdentityProviderError>;
 
-    /// Find federated user by IDP and Unique ID.
     async fn find_federated_user<'a>(
         &self,
         state: &ServiceState,
@@ -62,56 +53,50 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         unique_id: &'a str,
     ) -> Result<Option<UserResponse>, IdentityProviderError>;
 
-    /// Create user.
     async fn create_user(
         &self,
         state: &ServiceState,
         user: UserCreate,
     ) -> Result<UserResponse, IdentityProviderError>;
 
-    /// Delete user.
     async fn delete_user<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 
-    /// List groups.
     async fn list_groups(
         &self,
         state: &ServiceState,
         params: &GroupListParameters,
-    ) -> Result<Vec<Group>, IdentityProviderError>;
+    ) -> Result<impl IntoIterator<Item = Group>, IdentityProviderError>;
 
-    /// Get single group by ID.
     async fn get_group<'a>(
         &self,
         state: &ServiceState,
         group_id: &'a str,
     ) -> Result<Option<Group>, IdentityProviderError>;
 
-    /// Create group.
     async fn create_group(
         &self,
         state: &ServiceState,
         group: GroupCreate,
     ) -> Result<Group, IdentityProviderError>;
 
-    /// Delete group by ID.
     async fn delete_group<'a>(
         &self,
         state: &ServiceState,
         group_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 
-    /// List groups a user is member of.
+    /// List groups the user is a member of.
     async fn list_groups_of_user<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
-    ) -> Result<Vec<Group>, IdentityProviderError>;
+    ) -> Result<impl IntoIterator<Item = Group>, IdentityProviderError>;
 
-    /// Add the user to the group.
+    /// Add the user to the single group.
     async fn add_user_to_group<'a>(
         &self,
         state: &ServiceState,
@@ -119,14 +104,14 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         group_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Add user group membership relations.
+    /// Add user group memberships as specified by (uid, gid) tuples.
     async fn add_users_to_groups<'a>(
         &self,
         state: &ServiceState,
         memberships: Vec<(&'a str, &'a str)>,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Remove the user from the group.
+    /// Remove the user from the single group.
     async fn remove_user_from_group<'a>(
         &self,
         state: &ServiceState,
@@ -134,7 +119,7 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         group_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Remove the user from multiple groups.
+    /// Remove the user from specified groups.
     async fn remove_user_from_groups<'a>(
         &self,
         state: &ServiceState,
@@ -142,7 +127,7 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         group_ids: HashSet<&'a str>,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Set group memberships for the user.
+    /// Set group memberships of the user.
     async fn set_user_groups<'a>(
         &self,
         state: &ServiceState,
@@ -150,12 +135,11 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         group_ids: HashSet<&'a str>,
     ) -> Result<(), IdentityProviderError>;
 
-    /// List user passkeys.
     async fn list_user_webauthn_credentials<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
-    ) -> Result<Vec<Passkey>, IdentityProviderError>;
+    ) -> Result<impl IntoIterator<Item = Passkey>, IdentityProviderError>;
 
     /// Create passkey.
     async fn create_user_webauthn_credential<'a>(
@@ -166,49 +150,43 @@ pub trait IdentityBackend: DynClone + Send + Sync + std::fmt::Debug {
         description: Option<&'a str>,
     ) -> Result<WebauthnCredential, IdentityProviderError>;
 
-    /// Save passkey registration state.
-    async fn create_user_webauthn_credential_registration_state<'a>(
+    async fn save_user_webauthn_credential_registration_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
         state: PasskeyRegistration,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Save passkey auth state.
-    async fn create_user_webauthn_credential_authentication_state<'a>(
+    async fn save_user_webauthn_credential_authentication_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
         state: PasskeyAuthentication,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Get passkey registration state.
     async fn get_user_webauthn_credential_registration_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<Option<PasskeyRegistration>, IdentityProviderError>;
 
-    /// Get passkey authentication state.
     async fn get_user_webauthn_credential_authentication_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<Option<PasskeyAuthentication>, IdentityProviderError>;
 
-    /// Delete passkey registration state of a user.
+    /// Delete passkey registration state of a user
     async fn delete_user_webauthn_credential_registration_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 
-    /// Delete passkey authentication state of a user.
+    /// Delete passkey registration state of a user
     async fn delete_user_webauthn_credential_authentication_state<'a>(
         &self,
         state: &ServiceState,
         user_id: &'a str,
     ) -> Result<(), IdentityProviderError>;
 }
-
-dyn_clone::clone_trait_object!(IdentityBackend);
